Refactor config loaders and domain core structure
Renamed Elephlink to Core in the domain layer and moved configuration loader classes to a unified 'infra/config/loaders' package.
Updated imports and usages throughout the codebase to reflect these changes.
