+++ src/main/java/com/roomelephant/elephlink/Main.java
@@ -7,17 +7,17 @@ import static com.roomelephant.elephlink.infra.config.ConfigurationProperties.RE
 import com.roomelephant.elephlink.adapters.cloudflare.CloudFlareServiceImpl;
 import com.roomelephant.elephlink.adapters.ipservice.IpServiceImpl;
 import com.roomelephant.elephlink.domain.CloudFlareService;
-import com.roomelephant.elephlink.domain.Elephlink;
+import com.roomelephant.elephlink.domain.Core;
 import com.roomelephant.elephlink.domain.IpService;
 import com.roomelephant.elephlink.domain.TaskManager;
 import com.roomelephant.elephlink.domain.model.AuthConfig;
 import com.roomelephant.elephlink.domain.model.DnsRecordsConfig;
 import com.roomelephant.elephlink.domain.model.IpServiceConfig;
 import com.roomelephant.elephlink.infra.TaskManagerImpl;
-import com.roomelephant.elephlink.infra.config.ConfigLoader;
-import com.roomelephant.elephlink.infra.config.auth.AuthConfigurationLoader;
-import com.roomelephant.elephlink.infra.config.dnsrecords.RecordsConfigurationLoader;
-import com.roomelephant.elephlink.infra.config.ipservice.IpServiceConfigurationLoader;
+import com.roomelephant.elephlink.infra.ConfigLoader;
+import com.roomelephant.elephlink.infra.config.loaders.AuthConfigurationLoader;
+import com.roomelephant.elephlink.infra.config.loaders.RecordsConfigurationLoader;
+import com.roomelephant.elephlink.infra.config.loaders.IpServiceConfigurationLoader;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.function.Function;
@@ -30,7 +30,7 @@ public class Main {
     log.info("Starting Elephlink ...");
     Map<String, String> parameters = parser.apply(args);
 
-    Elephlink elephlink;
+    Core core;
     try {
       ConfigLoader<AuthConfig> authLoader = new AuthConfigurationLoader();
       AuthConfig authConfig = authLoader.load(parameters.get(AUTH_CONFIGURATION_FILE.key()));
@@ -44,7 +44,7 @@ public class Main {
       IpServiceConfig ipConfig = ipServicesLoader.load(parameters.get(IP_LIST_CONFIGURATION_FILE.key()));
       IpService ipServiceImpl = new IpServiceImpl(ipConfig);
 
-      elephlink = new Elephlink(dnsRecordsConfig, cloudFlareServiceImpl, ipServiceImpl, taskManager);
+      core = new Core(dnsRecordsConfig, cloudFlareServiceImpl, ipServiceImpl, taskManager);
     } catch (Exception e) {
       log.error("Validate your configuration. {}", e.getMessage());
       System.exit(1);
@@ -52,13 +52,13 @@ public class Main {
     }
 
     try {
-      elephlink.validateConfigurations();
+      core.validateConfigurations();
     } catch (Exception e) {
       log.error("Invalid configuration. {}", e.getMessage());
       System.exit(1);
       return;
     }
-    elephlink.start();
+    core.start();
   }
 
   private static final Function<String[], Map<String, String>> parser = parameters -> Arrays.stream(parameters)
