+++ src/main/java/com/roomelephant/elephlink/infra/config/loaders/AuthConfigurationLoader.java
@@ -1,13 +1,11 @@
-package com.roomelephant.elephlink.infra.config.auth;
+package com.roomelephant.elephlink.infra.config.loaders;
 
 import static com.roomelephant.elephlink.infra.config.ConfigurationFiles.AUTH_FILE;
-import static com.roomelephant.elephlink.infra.config.auth.AuthConfigurationLoader.AuthProperties.AUTH_EMAIL;
-import static com.roomelephant.elephlink.infra.config.auth.AuthConfigurationLoader.AuthProperties.AUTH_KEY;
-import static com.roomelephant.elephlink.infra.config.auth.AuthConfigurationLoader.AuthProperties.AUTH_METHOD;
-import static com.roomelephant.elephlink.infra.config.auth.AuthConfigurationLoader.AuthProperties.ZONE_IDENTIFIER;
+import static com.roomelephant.elephlink.infra.config.loaders.AuthConfigurationLoader.AuthProperties.AUTH_EMAIL;
+import static com.roomelephant.elephlink.infra.config.loaders.AuthConfigurationLoader.AuthProperties.AUTH_KEY;
+import static com.roomelephant.elephlink.infra.config.loaders.AuthConfigurationLoader.AuthProperties.ZONE_IDENTIFIER;
 
 import com.roomelephant.elephlink.domain.model.AuthConfig;
-import com.roomelephant.elephlink.infra.config.BaseConfigurationLoader;
 import java.util.Map;
 
 public class AuthConfigurationLoader extends BaseConfigurationLoader<AuthConfig> {
@@ -15,28 +13,14 @@ public class AuthConfigurationLoader extends BaseConfigurationLoader<AuthConfig>
   @Override
   protected AuthConfig convert(Map<String, Object> ymlConfig) {
     String email = getAndValidateString(ymlConfig, AUTH_EMAIL.key());
-    AuthConfig.Method method = getAndValidateEnumMethod(ymlConfig);
     String key = getAndValidateString(ymlConfig, AUTH_KEY.key());
     String zoneIdentifier = getAndValidateString(ymlConfig, ZONE_IDENTIFIER.key());
 
     return AuthConfig.builder()
         .email(email)
-        .method(method)
         .key(key)
         .zoneIdentifier(zoneIdentifier)
         .build();
-  }
-
-  private AuthConfig.Method getAndValidateEnumMethod(Map<String, Object> ymlConfig) {
-    String methodRaw = getAndValidateString(ymlConfig, AUTH_METHOD.key());
-    AuthConfig.Method method;
-    try {
-      method = AuthConfig.Method.fromString(methodRaw);
-    } catch (Exception e) {
-      throw new IllegalArgumentException(
-          "Invalid auth configuration. " + methodRaw + " is not recognized. Use 'token' or 'global'.");
-    }
-    return method;
   }
 
   @Override
@@ -54,7 +38,6 @@ public class AuthConfigurationLoader extends BaseConfigurationLoader<AuthConfig>
 
   enum AuthProperties {
     AUTH_EMAIL("email"),
-    AUTH_METHOD("method"),
     AUTH_KEY("key"),
     ZONE_IDENTIFIER("zoneIdentifier");
